refactor(event): modularize event router
