+++ app/(main)/@form/communities/[slug]/view/subscribe/page.tsx
@@ -11,7 +11,7 @@ export default async function SubscribePage({ params }: SubscribePageProps) {
 	const api = await getAPI()
 
 	try {
-		const community = await api.community.get({ slug })
+		const community = await api.community.get.enhanced({ slug })
 		if (!community) {
 			return notFound()
 		}
