+++ server/api/routers/index.ts
@@ -1,7 +1,7 @@
-export * from './category'
-export * from './community'
-export * from './event'
-export * from './image'
-export * from './location'
-export * from './rsvp'
-export * from './user'
+export * from '@/server/api/routers/category'
+export * from '@/server/api/routers/community'
+export * from '@/server/api/routers/event'
+export * from '@/server/api/routers/image'
+export * from '@/server/api/routers/location'
+export * from '@/server/api/routers/rsvp'
+export * from '@/server/api/routers/user'
