+++ server/api/shared/includes.ts
@@ -1,0 +1,1 @@
+export const prismaIncludes = {}
