+++ server/api/shared/types.ts
@@ -1,0 +1,4 @@
+import type { z } from 'zod'
+import type { PaginationSchema } from './schemas'
+
+export type PaginationInput = z.infer<typeof PaginationSchema>
